pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cran-ai"
version = "0.1.0"
description = "6G-ready AI-native Cognitive Relay Framework (DF/CF, robust CSI, multi-objective, hybrid selection, generalization, deploy)."
readme = "README.md"
requires-python = ">=3.11,<3.12"
authors = [{name = "CRAN-AI Authors"}]
license = {text = "Proprietary"}  # change later if needed
dependencies = [
  "numpy==2.0.2",
  "scipy==1.14.1",
  "pandas==2.2.3",
  "PyYAML==6.0.2",
  "scikit-learn==1.5.2",
  "matplotlib==3.10.8",
  "seaborn==0.13.2",
  "tqdm==4.67.1",
  "rich==13.9.4",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "jupyter==1.1.1",
  "ipykernel==6.29.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
